requirements-dev.txt
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-sugar==0.9.7  # Better test output formatting
pytest-randomly==3.12.0  # Random test ordering

# Code Quality
black==23.7.0
ruff==0.0.285
mypy==1.4.1

# Development
ipython==8.14.0

# Integration test dependencies
requests==2.31.0
websocket-client==1.6.1